Commit stable-4.2

+++ src/iostream.c
@@ -98,14 +98,55 @@ Obj FuncIS_BLOCKED_IOSTREAM( Obj self, Obj stream )
 
 #else
 
+#if HAVE_UNISTD_H
 #include <unistd.h>
+#endif
+
+#if HAVE_ERRNO_H
 #include <errno.h>
+#endif
+
+#if HAVE_SIGNAL_H
 #include <signal.h>
+#endif
+
+#if HAVE_FCNTL_H
 #include <fcntl.h>
+#endif
+
+#if HAVE_TERMIOS_H
 #include <termios.h>
+#endif
+
+#if HAVE_SYS_TYPES_H
 #include <sys/types.h>
+#endif
+
+#if HAVE_SYS_WAIT_H
 #include <sys/wait.h>
+#endif
+
+#if HAVE_ASSERT_H
 #include <assert.h>
+#else
+#ifdef NDEBUG
+#define assert( a )
+#else
+#define assert( a ) do if (!(a)) {fprintf(stderr,"Assertion failed at line %d file %s\n",__LINE__,__FILE__); abort();} while (0)
+#endif
+#endif
+
+#if HAVE_SYS_STAT_H
+#include <sys/stat.h>
+#endif
+
+#if HAVE_STDLIB_H
+#include <stdlib.h>
+#endif
+
+#if HAVE_STRING_H
+#include <string.h>
+#endif
 
 typedef struct {
   int childPID;    /* Also used as a link to make a linked free list */
@@ -260,6 +301,16 @@ void KillChild (UInt stream)
 
 static UInt GetMasterPty ( int * pty, Char * ttyname, Char *ptyname )
 {
+#if HAVE_GETPT && HAVE_PTSNAME_R
+  if ((*pty = getpt()) > 0 )
+    {
+      if (grantpt(*pty) || unlockpt(*pty))
+	return 1;
+      ptsname_r(*pty, ttyname, 32); 
+      return 0;
+    }
+  return 1;
+#else
 #   ifdef att
         if ( (*pty = open( "/dev/ptmx", O_RDWR )) < 0 )
             return 1;
@@ -328,6 +379,7 @@ static UInt GetMasterPty ( int * pty, Char * ttyname, Char *ptyname )
 #   endif
 #   endif
 #   endif
+#endif
 }
 
 
@@ -361,6 +413,17 @@ RETSIGTYPE ChildStatusChanged( int whichsig )
 	    }
 	}
     }
+  /* Collect up any other zombie children */
+  do {
+   retcode = waitpid( -1, &status, WNOHANG);
+   if (retcode == -1 && errno != ECHILD)
+     Pr("#E Unexpected waitpid error %d\n",errno, 0);
+#if 0
+   else if (retcode != 0)
+     Pr"Reaped unexpected child %d\n",retcode, 0L);
+#endif
+ } while (retcode != 0 && retcode != -1);
+  
   signal(SIGCHLD, ChildStatusChanged);
 }
 
@@ -482,7 +545,7 @@ Int StartChildProcess ( Char *dir, Char *prg, Char *args[] )
 	if ( PtyIOStreams[stream].childPID == 0 )
 	  {
 	    /* Set up the child */
-	    
+            close(PtyIOStreams[stream].ptyFD);
 	    if ( dup2( slave, 0 ) == -1)
 	      _exit(-1);
 	    fcntl( 0, F_SETFD, 0 );
@@ -513,7 +576,7 @@ Int StartChildProcess ( Char *dir, Char *prg, Char *args[] )
 	    Pr( "Panic: cannot fork to subprocess.\n", 0, 0);
 	    goto cleanup;
 	  }
-
+        close(slave);
 	
 	
 	return stream;
@@ -636,8 +699,10 @@ Obj FuncCLOSE_PTY_IOSTREAM( Obj self, Obj stream )
   PtyIOStreams[pty].inuse = 0;
   
   /* Close down the child */
+  retcode = close(PtyIOStreams[pty].ptyFD);
+  if (retcode)
+    Pr("Strange close return code %d\n",retcode, 0);
   kill(PtyIOStreams[pty].childPID, SIGTERM);
-  close(PtyIOStreams[pty].ptyFD);
   retcode = waitpid(PtyIOStreams[pty].childPID, &status, 0);
   FreeStream(pty);
   return 0;
@@ -712,12 +777,18 @@ static Int InitKernel(
       StructInitInfo * module )
 {
 #if SYS_MAC_MWC || SYS_MAC_MPW
+
+  InitHdlrFuncsFromTable( GVarFuncs );
+
 #else
 
   UInt i;
   PtyIOStreams[0].childPID = -1;
   for (i = 1; i < MAX_PTYS; i++)
-    PtyIOStreams[i].childPID = i-1;
+    {
+      PtyIOStreams[i].childPID = i-1;
+      PtyIOStreams[i].inuse = 0;
+    }
   FreePtyIOStreams = MAX_PTYS-1;
 
   /* init filters and functions                                          */
